Reduce allocations and calls to Write() in the RPCv2 server
Right now the RPCv2 server performs two calls to Write() for every
successful RPC response. This change brings that back to one by making
sure that the full reply is stored in a contiguous buffer. We let the
Service write its response into a bytes.Buffer, but make sure that
enough leading space is present for the header.

Because the header's size is in theory variable, we make sure to memoize
the size of the largest header that's been observed thus far. This means
that we can allocate an appropriately sized buffer in the common case,
falling back to copying data into a new buffer when it needs to grow.

We also do the same thing with the return values. This ensures that the
RPCv2 server only needs to allocate a single buffer for every RPC
processed.

As all replies are now done through a single Write(), we no longer need
to use a separate mutex. This is safe, as documented in the "net"
package:

    Multiple goroutines may invoke methods on a Conn simultaneously.