Add support for authentication
This change adds a new Authenticator interface that is called for every
incoming RPC. The Authenticator may inspect the incoming credentials and
verifier, and may return a verifier that is attached to the response.

No implementations are provided as part of this repository. We will add
one to bb-remote-execution for RFC 5531 appendix A system
authentication, because we'd like to use some of Buildbarn's helper
types for that.